requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# Data Processing
datasets>=2.12.0  # Optional: for HuggingFace datasets
tokenizers>=0.13.0  # Optional: for advanced tokenization

# Utilities
einops>=0.6.1

